+++ src/api/embed_helpers.cc
@@ -68,7 +68,7 @@ Maybe<int> SpinEventLoop(Environment* env) {
   env->set_snapshot_serialize_callback(Local<Function>());
 
   env->PrintInfoForSnapshotIfDebug();
-  env->VerifyNoStrongBaseObjects();
+  env->principal_realm()->VerifyNoStrongBaseObjects();
   return EmitProcessExit(env);
 }
