+++ test/cctest/test_base_object_ptr.cc
@@ -1,6 +1,7 @@
+#include "base_object-inl.h"
 #include "gtest/gtest.h"
 #include "node.h"
-#include "base_object-inl.h"
+#include "node_realm-inl.h"
 #include "node_test_fixture.h"
 
 using node::BaseObject;
@@ -9,6 +10,7 @@ using node::BaseObjectWeakPtr;
 using node::Environment;
 using node::MakeBaseObject;
 using node::MakeDetachedBaseObject;
+using node::Realm;
 using v8::HandleScope;
 using v8::Isolate;
 using v8::Local;
@@ -46,13 +48,14 @@ TEST_F(BaseObjectPtrTest, ScopedDetached) {
   const Argv argv;
   Env env_{handle_scope, argv};
   Environment* env = *env_;
+  Realm* realm = env->principal_realm();
 
-  EXPECT_EQ(env->base_object_created_after_bootstrap(), 0);
+  EXPECT_EQ(realm->base_object_created_after_bootstrap(), 0);
   {
     BaseObjectPtr<DummyBaseObject> ptr = DummyBaseObject::NewDetached(env);
-    EXPECT_EQ(env->base_object_created_after_bootstrap(), 1);
+    EXPECT_EQ(realm->base_object_created_after_bootstrap(), 1);
   }
-  EXPECT_EQ(env->base_object_created_after_bootstrap(), 0);
+  EXPECT_EQ(realm->base_object_created_after_bootstrap(), 0);
 }
 
 TEST_F(BaseObjectPtrTest, ScopedDetachedWithWeak) {
@@ -60,17 +63,18 @@ TEST_F(BaseObjectPtrTest, ScopedDetachedWithWeak) {
   const Argv argv;
   Env env_{handle_scope, argv};
   Environment* env = *env_;
+  Realm* realm = env->principal_realm();
 
   BaseObjectWeakPtr<DummyBaseObject> weak_ptr;
 
-  EXPECT_EQ(env->base_object_created_after_bootstrap(), 0);
+  EXPECT_EQ(realm->base_object_created_after_bootstrap(), 0);
   {
     BaseObjectPtr<DummyBaseObject> ptr = DummyBaseObject::NewDetached(env);
     weak_ptr = ptr;
-    EXPECT_EQ(env->base_object_created_after_bootstrap(), 1);
+    EXPECT_EQ(realm->base_object_created_after_bootstrap(), 1);
   }
   EXPECT_EQ(weak_ptr.get(), nullptr);
-  EXPECT_EQ(env->base_object_created_after_bootstrap(), 0);
+  EXPECT_EQ(realm->base_object_created_after_bootstrap(), 0);
 }
 
 TEST_F(BaseObjectPtrTest, Undetached) {
@@ -78,16 +82,17 @@ TEST_F(BaseObjectPtrTest, Undetached) {
   const Argv argv;
   Env env_{handle_scope, argv};
   Environment* env = *env_;
+  Realm* realm = env->principal_realm();
 
   node::AddEnvironmentCleanupHook(
       isolate_,
       [](void* arg) {
-        EXPECT_EQ(static_cast<Environment*>(arg)->base_object_count(), 0);
+        EXPECT_EQ(static_cast<Realm*>(arg)->base_object_count(), 0);
       },
-      env);
+      realm);
 
   BaseObjectPtr<DummyBaseObject> ptr = DummyBaseObject::New(env);
-  EXPECT_EQ(env->base_object_created_after_bootstrap(), 1);
+  EXPECT_EQ(realm->base_object_created_after_bootstrap(), 1);
 }
 
 TEST_F(BaseObjectPtrTest, GCWeak) {
@@ -95,6 +100,7 @@ TEST_F(BaseObjectPtrTest, GCWeak) {
   const Argv argv;
   Env env_{handle_scope, argv};
   Environment* env = *env_;
+  Realm* realm = env->principal_realm();
 
   BaseObjectWeakPtr<DummyBaseObject> weak_ptr;
 
@@ -104,21 +110,21 @@ TEST_F(BaseObjectPtrTest, GCWeak) {
     weak_ptr = ptr;
     ptr->MakeWeak();
 
-    EXPECT_EQ(env->base_object_created_after_bootstrap(), 1);
+    EXPECT_EQ(realm->base_object_created_after_bootstrap(), 1);
     EXPECT_EQ(weak_ptr.get(), ptr.get());
     EXPECT_EQ(weak_ptr->persistent().IsWeak(), false);
 
     ptr.reset();
   }
 
-  EXPECT_EQ(env->base_object_created_after_bootstrap(), 1);
+  EXPECT_EQ(realm->base_object_created_after_bootstrap(), 1);
   EXPECT_NE(weak_ptr.get(), nullptr);
   EXPECT_EQ(weak_ptr->persistent().IsWeak(), true);
 
   v8::V8::SetFlagsFromString("--expose-gc");
   isolate_->RequestGarbageCollectionForTesting(Isolate::kFullGarbageCollection);
 
-  EXPECT_EQ(env->base_object_created_after_bootstrap(), 0);
+  EXPECT_EQ(realm->base_object_created_after_bootstrap(), 0);
   EXPECT_EQ(weak_ptr.get(), nullptr);
 }
 
@@ -127,9 +133,10 @@ TEST_F(BaseObjectPtrTest, Moveable) {
   const Argv argv;
   Env env_{handle_scope, argv};
   Environment* env = *env_;
+  Realm* realm = env->principal_realm();
 
   BaseObjectPtr<DummyBaseObject> ptr = DummyBaseObject::NewDetached(env);
-  EXPECT_EQ(env->base_object_created_after_bootstrap(), 1);
+  EXPECT_EQ(realm->base_object_created_after_bootstrap(), 1);
   BaseObjectWeakPtr<DummyBaseObject> weak_ptr { ptr };
   EXPECT_EQ(weak_ptr.get(), ptr.get());
 
@@ -140,12 +147,12 @@ TEST_F(BaseObjectPtrTest, Moveable) {
   BaseObjectWeakPtr<DummyBaseObject> weak_ptr2 = std::move(weak_ptr);
   EXPECT_EQ(weak_ptr2.get(), ptr2.get());
   EXPECT_EQ(weak_ptr.get(), nullptr);
-  EXPECT_EQ(env->base_object_created_after_bootstrap(), 1);
+  EXPECT_EQ(realm->base_object_created_after_bootstrap(), 1);
 
   ptr2.reset();
 
   EXPECT_EQ(weak_ptr2.get(), nullptr);
-  EXPECT_EQ(env->base_object_created_after_bootstrap(), 0);
+  EXPECT_EQ(realm->base_object_created_after_bootstrap(), 0);
 }
 
 TEST_F(BaseObjectPtrTest, NestedClasses) {
@@ -165,18 +172,19 @@ TEST_F(BaseObjectPtrTest, NestedClasses) {
   const Argv argv;
   Env env_{handle_scope, argv};
   Environment* env = *env_;
+  Realm* realm = env->principal_realm();
 
   node::AddEnvironmentCleanupHook(
       isolate_,
       [](void* arg) {
-        EXPECT_EQ(static_cast<Environment*>(arg)->base_object_count(), 0);
+        EXPECT_EQ(static_cast<Realm*>(arg)->base_object_count(), 0);
       },
-      env);
+      realm);
 
   ObjectWithPtr* obj =
       new ObjectWithPtr(env, DummyBaseObject::MakeJSObject(env));
   obj->ptr1 = DummyBaseObject::NewDetached(env);
   obj->ptr2 = DummyBaseObject::New(env);
 
-  EXPECT_EQ(env->base_object_created_after_bootstrap(), 3);
+  EXPECT_EQ(realm->base_object_created_after_bootstrap(), 3);
 }
