+++ test/pummel/test-heapdump-env.js
@@ -21,11 +21,16 @@ validateSnapshotNodes('Node / Environment', [{
   ]
 }]);
 
-validateSnapshotNodes('Node / CleanupQueue', [{
-  children: [
-    { node_name: 'Node / ContextifyScript' },
-  ]
-}]);
+validateSnapshotNodes('Node / CleanupQueue', [
+  // The first one is the cleanup_queue of the Environment.
+  {},
+  // The second one is the cleanup_queue of the principal realm.
+  {
+    children: [
+      { node_name: 'Node / ContextifyScript' },
+    ]
+  },
+]);
 
 validateSnapshotNodes('Node / Realm', [{
   children: [
